requirements.txt
django==4.1.4
Pillow==9.5.0
azure-storage-blob==12.17.0
django-storages==1.13.2
django-filter==25.1
django-sslserver
django-cors-headers
djangorestframework
python-dotenv
psycopg2
nh3